chore: api 관련 코드 정리
