+++ src/app/entry-meeting/_components/LinkInput.tsx
@@ -3,7 +3,7 @@ import { useForm } from 'react-hook-form'
 import Image from 'next/image'
 import { useRouter } from 'next/navigation'
 import { z } from 'zod'
-import { useCheckMeetingLink } from '@/apis/queries/meetingQueries'
+import { useCheckMeetingLink } from '@/apis/meetingApi'
 import { Button } from '@/components/Button'
 import Callout from '@/components/Callout'
 import { TextInput } from '@/components/Inputs/TextInput'
