+++ src/app/entry-meeting/_components/MeetingInfo.tsx
@@ -2,8 +2,8 @@ import { useCallback, useEffect } from 'react'
 import dayjs from 'dayjs'
 import Image from 'next/image'
 import { useRouter } from 'next/navigation'
-import { useCheckMeetingLink } from '@/apis/queries/meetingQueries'
-import { useGetParticipantsMe } from '@/apis/queries/participantsQueries'
+import { useCheckMeetingLink } from '@/apis/meetingApi'
+import { useGetParticipantsMe } from '@/apis/participantsApi'
 import { Button } from '@/components/Button'
 import Loading from '@/components/Loading'
 import { IMAGE_BASE_URL } from '@/constant/base_url'
