+++ src/app/entry-meeting/_components/NicknameInput.tsx
@@ -3,7 +3,7 @@ import { useForm } from 'react-hook-form'
 import { zodResolver } from '@hookform/resolvers/zod'
 import Image from 'next/image'
 import { z } from 'zod'
-import { useCheckNickname, useJoinMeeting } from '@/apis/queries/meetingQueries'
+import { useCheckNickname, useJoinMeeting } from '@/apis/meetingApi'
 import { Button } from '@/components/Button'
 import { TextInput } from '@/components/Inputs/TextInput'
 import useDebounce from '@/hooks/useDebounce'
