+++ src/app/manage-meeting/page.tsx
@@ -6,7 +6,7 @@ import { zodResolver } from '@hookform/resolvers/zod'
 import Image from 'next/image'
 import { useRouter } from 'next/navigation'
 import { z } from 'zod'
-import { useModifyMeeting } from '@/apis/queries/meetingQueries'
+import { useModifyMeeting } from '@/apis/meetingApi'
 import { Button } from '@/components/Button'
 import ColorPicker from '@/components/ColorPicker'
 import { TextInput } from '@/components/Inputs/TextInput'
