+++ src/app/manage-mission/page.tsx
@@ -3,10 +3,7 @@
 import { useEffect, useState } from 'react'
 import Image from 'next/image'
 import { useRouter } from 'next/navigation'
-import {
-  useDeleteMission,
-  useGetLeaderMission,
-} from '@/apis/queries/missionQuries'
+import { useDeleteMission, useGetLeaderMission } from '@/apis/missionApi'
 import { Button } from '@/components/Button'
 import Loading from '@/components/Loading'
 import useMeetingStore from '@/stores/useMeetingStore'
