+++ src/app/meeting/info/_components/MeetingDetail.tsx
@@ -2,7 +2,7 @@ import { useCallback, useEffect, useRef, useState } from 'react'
 import dayjs from 'dayjs'
 import Image from 'next/image'
 import Link from 'next/link'
-import { useParticipants } from '@/apis/queries/participantsQueries'
+import { useParticipants } from '@/apis/participantsApi'
 import Loading from '@/components/Loading'
 import useMeetingStore from '@/stores/useMeetingStore'
 import useUserStore from '@/stores/useUserStore'
