+++ src/app/meeting/photo-capture/_components/PhotoView.tsx
@@ -2,8 +2,8 @@ import { useEffect } from 'react'
 import { format } from 'date-fns/format'
 import Image from 'next/image'
 import { useRouter } from 'next/navigation'
-import { useGetParticipantsMe } from '@/apis/queries/participantsQueries'
-import { useUploadSnap } from '@/apis/queries/snapQueries'
+import { useGetParticipantsMe } from '@/apis/participantsApi'
+import { useUploadSnap } from '@/apis/snapApi'
 import Refresh from '@/assets/Refresh.svg'
 import { Button } from '@/components/Button'
 import Tooltip from '@/components/Tooltip'
