+++ src/app/meeting/photo/[photoId]/page.tsx
@@ -2,7 +2,7 @@
 
 import Image from 'next/image'
 import { useRouter } from 'next/navigation'
-import { useGetSnapDetail } from '@/apis/queries/snapQueries'
+import { useGetSnapDetail } from '@/apis/snapApi'
 import AuthGuard from '@/app/AuthGuard'
 import { Button } from '@/components/Button'
 import Loading from '@/components/Loading'
